Fix html, clean up a bit, split furnace display into 3 subcomponents for easier understanding
